improved simulation saving switched to floats and removed streams

+++ src/main/java/biology/Brain.java
@@ -8,11 +8,11 @@ import core.Simulation;
 public interface Brain extends Serializable
 {
 	void tick(Protozoa p);
-	double turn(Protozoa p);
-	double speed(Protozoa p);
+	float turn(Protozoa p);
+	float speed(Protozoa p);
 	boolean wantToAttack(Protozoa p);
 	boolean wantToMateWith(Protozoa p);
-	double energyConsumption();
+	float energyConsumption();
 	
 	Brain RANDOM = new Brain()
 	{
@@ -22,16 +22,16 @@ public interface Brain extends Serializable
 		public void tick(Protozoa p) {}
 
 		@Override
-		public double turn(Protozoa p)
+		public float turn(Protozoa p)
 		{
-			double x = 2*Simulation.RANDOM.nextDouble() - 1;
-			double t = Math.toRadians(35);
+			float x = (float) (2*Simulation.RANDOM.nextDouble() - 1);
+			float t = (float) Math.toRadians(35);
 			return t * x;
 		}
 
 		@Override
-		public double speed(Protozoa p) {
-			return Simulation.RANDOM.nextDouble() * Settings.maxVel;
+		public float speed(Protozoa p) {
+			return (float) (Simulation.RANDOM.nextDouble() * Settings.maxVel);
 		}
 
 		@Override
@@ -46,7 +46,7 @@ public interface Brain extends Serializable
 		}
 
 		@Override
-		public double energyConsumption() {
+		public float energyConsumption() {
 			return 0;
 		}
